# docs/sdlc/phase-4-design/openapi.yaml
openapi: 3.0.3
info:
  title: Automated Regulatory Filing Module API
  version: 0.1.0-mvp
  description: >
    Generated by L1-design-api-spec (Phase 4). Correlation ID C885C23C-949E-440C-8B0E-04FDD166A559.
    MVP scope only (E1.F1, E1.F2, E1.F3, E1.F4, E2.F2, E2.F3). Auth is a mocked SSO bearer token — no real corporate SSO
    integration in this demo. EDGAR submission is mocked — no real external filing occurs.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
paths:
  /demo/simulate-trade:
    post:
      summary: Demo-only — simulate an EquatePlus transaction execution event (stands in for FR-001's real event source)
      operationId: simulateTrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [executiveId, transactionCode, shares, pricePerShare]
              properties:
                executiveId: { type: string }
                transactionCode: { type: string, enum: [S, P, A, D], description: "SEC transaction code (Sale, Purchase, Grant/Award, Disposition)" }
                shares: { type: number }
                pricePerShare: { type: number }
      responses:
        "201":
          description: Transaction accepted; a Filing is generated asynchronously within 5 minutes (FR-003) — MVP generates it synchronously for demo purposes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Filing" }
  /filings:
    get:
      summary: List filings, filterable by status/executive (FR-013, FR-015)
      operationId: listFilings
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/FilingStatus" }
        - name: executiveId
          in: query
          schema: { type: string }
        - name: search
          in: query
          description: Free-text match against executive display name or issuer (used by the Compliance Dashboard's search box)
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Filing" }
  /filings/{id}:
    get:
      summary: Get one filing with its fields (FR-009)
      operationId: getFiling
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Filing" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Edit filing fields; triggers re-validation (FR-010, FR-004)
      operationId: updateFiling
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fields:
                  type: object
                  additionalProperties: true
      responses:
        "200": { description: Updated and re-validated, content: { application/json: { schema: { $ref: "#/components/schemas/Filing" } } } }
        "422": { description: Validation failed, content: { application/json: { schema: { $ref: "#/components/schemas/ValidationErrorResponse" } } } }
  /filings/{id}/approve:
    post:
      summary: Approve a validated filing; triggers mocked EDGAR submission connector (FR-011, FR-020)
      operationId: approveFiling
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Filing approved and (mock-)submitted, content: { application/json: { schema: { $ref: "#/components/schemas/Filing" } } } }
        "409": { description: Filing has not passed validation; approval blocked }
  /filings/{id}/audit-log:
    get:
      summary: Audit trail for one filing (FR-012)
      operationId: getFilingAuditLog
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditLogEntry" }
  /notifications:
    get:
      summary: List notifications for the current user (FR-006, FR-007, FR-008)
      operationId: listNotifications
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, description: "Mocked SSO token — not a real corporate SSO integration in this MVP." }
  responses:
    NotFound:
      description: Not found
      content: { application/json: { schema: { type: object, properties: { message: { type: string } } } } }
  schemas:
    FilingStatus:
      type: string
      enum: [INCOMPLETE, UNDER_REVIEW, VALIDATED, SUBMITTED]
      description: >
        GENERATED and APPROVED are audit-log action names, not resting statuses — a filing
        transitions directly UNDER_REVIEW -> VALIDATED -> SUBMITTED (or INCOMPLETE if
        required data was missing at generation). See audit_log.action for the full action set.
    Filing:
      type: object
      properties:
        id: { type: string }
        formType: { type: string, enum: [FORM_4], description: "Modular per form-template-registry (E3.F3); MVP only registers FORM_4." }
        executiveId: { type: string }
        status: { $ref: "#/components/schemas/FilingStatus" }
        fields: { type: object, additionalProperties: true, description: "SEC Form 4 fields, e.g. issuer, reportingPerson, transactionDate, transactionCode, shares, pricePerShare." }
        deadlineAt: { type: string, format: date-time, description: "Statutory deadline (2 business days from transaction)." }
        edgarCutoffAt: { type: string, format: date-time, description: "Effective same-day EDGAR cutoff (5:30pm ET) per vision.md Regulatory Posture #2 — used for deadline-risk calculation, not deadlineAt directly." }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ValidationErrorResponse:
      type: object
      properties:
        filingId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    AuditLogEntry:
      type: object
      properties:
        id: { type: string }
        filingId: { type: string }
        action: { type: string, enum: [GENERATED, EDITED, APPROVED, SUBMITTED, FLAGGED_INCOMPLETE] }
        actorId: { type: string }
        timestamp: { type: string, format: date-time }
        detail: { type: string }
    Notification:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        type: { type: string, enum: [FORM_GENERATED, DEADLINE_RISK, SUBMITTED, INCOMPLETE_DATA] }
        filingId: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
        read: { type: boolean }